23.05.08 코드업 기초예제 11023번 실수를 입력받아 정수와 실수 부분으로 나누어 출력하기.

+++ src/Grass/Cu1023.java
@@ -11,23 +11,31 @@ public class Cu1023 { // 실수 1개 입력받아 부분별로 출력하기
 
         String i = numArr[0];
         String f = numArr[1];
+
 // 풀이 (1)
         int a = Integer.parseInt(i);
         int b = Integer.parseInt(f); // 실수 부분이지만 소수점 이하를 절삭하여 출력하기위해 int로 선언.
 
         System.out.println(a);
-        System.out.print((int)b);
+        System.out.print(b); // int b로 선언하면서 형 변환해줬으므로 그대로 출력
 
-//// 풀이 (2)
+// 풀이 (2)
+//       int a = Integer.parseInt(i);
+//       float b = Float.parseFloat(f); // 실수 부분이니까 실수로 형변환
+//
+//       System.out.println(a);
+//       System.out.print((int)b); // float으로 출력하면 '.0'이 출력되므로 소수점 이하를 절삭하기위해 정수로 형변환.
+
+// 풀이(3)
+//
 //        int a = Integer.parseInt(i);
-//        float b = Integer.parseInt(f);
+//        float b = Float.parseFloat(f); // 실수 부분이니까 실수로 형변환
 //
 //        System.out.println(a);
-//        System.out.print((int)b); // 소수점 이하를 절삭하기위해 int로 형변환.
-
+//        System.out.printf("%d", (int)b); // 여기서 int로 형변환했으므로 %d
 
     }
 }
 
-// parseInt : String 문자열로 된 숫자를 Int 등의  숫자타입으로 변환해줌(단, 문자열이 무조건 숫자로만 이루어져있어야함)
+// parseInt : String 문자열로 된 숫자를 Int 등의 숫자타입으로 변환해줌(단, 문자열이 무조건 숫자로만 이루어져있어야함)
 // Byte.parseByte(); short.ParseShort(); Long.parseLong(); Float.parseFloat(); 등등
